Apply patch for #5210
Fix a problem with the deletion of fai classes:
The query weither a class is associated to a profile matched
classes, which contained the name of a to-be-deleted class as a
substring of their own name.
The query was changed in a way that the list of associated classes
is compared exactly with the class to be deleted.

git-svn-id: https://oss.gonicus.de/repositories/gosa/branches/2.6-lhm@19542 594d385d-05f5-0310-b6e9-bd551577e9d8